provider/aws: availability zone data sources
In this case we provide both plural and singular data sources. The former
allows listing the available availability zones by name, to allow
use-cases such as "find me the first AZ available in my account", while
the latter allows retrieving data about a specific AZ.

As a helper for writing reusable modules, the AZ letter (without its
usual region name prefix) is exposed so that it can be used in
region-agnostic mappings where a different value is used per AZ, such as
for subnet numbering schemes.